revisi terakhir

+++ app/models/kategori_model.php
@@ -58,11 +58,41 @@ class kategori_model
 
     public function delete($id_kategori)
     {
-        $query = "DELETE FROM kategori WHERE id_kategori = :id_kategori";
+        $query = "SELECT id_artikel FROM kontributor WHERE id_kategori = :id_kategori";
         $this->db->query($query);
         $this->db->bind("id_kategori", $id_kategori);
         $this->db->execute();
+        $data['id_artikel'] = $this->db->resultSet();
 
-        return $this->db->rowCount();
+        // Delete related articles
+        if (!empty($data['id_artikel'])) {
+            $query = "DELETE FROM kontributor WHERE id_kategori = :id_kategori";
+            $this->db->query($query);
+            $this->db->bind("id_kategori", $id_kategori);
+            $this->db->execute();
+
+            if ($this->db->rowCount() > 0) {
+                $query = "DELETE FROM kategori WHERE id_kategori = :id_kategori";
+                $this->db->query($query);
+                $this->db->bind("id_kategori", $id_kategori);
+                $this->db->execute();
+                if ($this->db->rowCount() > 0) {
+                    foreach ($data['id_artikel'] as $row) {
+                        $query = "DELETE FROM artikel WHERE id_artikel = :id_artikel";
+                        $this->db->query($query);
+                        $this->db->bind("id_artikel", $row['id_artikel']);
+                        $this->db->execute();
+                        if ($this->db->rowCount() > 0) {
+                            $deletionsOccurred = true;
+                        }
+                    }
+                    if ($deletionsOccurred){
+                        return 1;
+                    }
+                }
+            } else {
+                return -1;
+            }
+        }
     }
 }
